+++ src/nsidc/metgen/metgen.py
@@ -347,7 +347,7 @@ def process(configuration: config.Config) -> None:
             premet_filename=premet_file,
             spatial_filename=spatial_file,
             reference_data_filename=reference_data_file,
-            data_reader=data_reader(configuration.auth_id, data_files),
+            data_reader=data_reader(configuration.auth_id, reference_data_file),
         )
         for name, reference_data_file, data_files, browse_files, premet_file, spatial_file in grouped_granule_files(
             configuration
@@ -360,16 +360,12 @@ def process(configuration: config.Config) -> None:
 
 
 def data_reader(
-    auth_id: str, data_files: set[str]
+    auth_id: str, data_file: str
 ) -> Callable[[str, str, str, config.Config], dict]:
     """
-    Determine which file reader to use for the given data files. This currently
-    is limited to handling one data file type (and one reader) per collection.
-    In a future issue, we may handle granules with multiple data file types per granule.
-    In that future work this needs to be refactored to handle this case.
+    Determine which file reader to use for the given data file.
     """
-    # Lookup based on an arbitrary data file in the set
-    _, extension = os.path.splitext(first(data_files))
+    _, extension = os.path.splitext(data_file)
 
     try:
         return registry.lookup(auth_id, extension)